# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AIA 论坛后端。",
        "title": "AIA 社团博客 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/articles": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "博客接口"
                ],
                "summary": "获取所有博客文章列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/article.Article"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "博客接口"
                ],
                "summary": "发布一篇 Markdown 博客",
                "parameters": [
                    {
                        "description": "文章数据",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/article.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/article.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": [
                    "博客接口"
                ],
                "summary": "获取特定 ID 的博客",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "博客ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/article.Article"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/article.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/article.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/assets": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "资源接口"
                ],
                "summary": "上传资源",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/asset.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资源接口"
                ],
                "summary": "上传资源文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/upload.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/upload.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "article.Article": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string",
                    "example": "张三"
                },
                "content": {
                    "type": "string",
                    "example": "# Markdown内容"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "标题五个字"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "article.CreateRequest": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "author": {
                    "type": "string",
                    "example": "社团负责人"
                },
                "content": {
                    "type": "string",
                    "example": "# 第一章：拒绝模糊逻辑\n这是一篇关于 Go 强类型的文章。"
                },
                "title": {
                    "type": "string",
                    "example": "关于后端严谨性的深度思考"
                }
            }
        },
        "article.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "参数校验失败：标题不能为空"
                }
            }
        },
        "asset.SuccessResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "upload.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "错误描述"
                }
            }
        },
        "upload.UploadResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "message": {
                    "type": "string",
                    "example": "上传成功"
                },
                "url": {
                    "type": "string",
                    "example": "/uploads/images/2026/03/25/uuid.png"
                }
            }
        }
    }
}